% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{convert_dir_to_df}
\alias{convert_dir_to_df}
\title{Convert Directory File Paths to a Data Frame}
\usage{
convert_dir_to_df(dir_path = find_root_directory())
}
\arguments{
\item{dir_path}{A character string specifying the path to the directory. Default should be the root directory.}
}
\value{
A data frame where each row represents a file. The columns represent
different levels of the directory structure.
}
\description{
This function lists all files in the specified directory recursively,
filters out paths without extensions, sorts them so that files with directory
names are pushed to the top, and then converts the sorted paths to a data frame
with levels representing the directory structure.
}
