% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{determine_modal_message}
\alias{determine_modal_message}
\title{Determine the Modal Message for Git Synchronization}
\usage{
determine_modal_message(
  ahead = gert::git_ahead_behind()$ahead,
  behind = gert::git_ahead_behind()$behind
)
}
\arguments{
\item{ahead}{Integer indicating how many commits the local repository
is ahead of the remote repository.}

\item{behind}{Integer indicating how many commits the local repository
is behind the remote repository.}
}
\value{
A character string with the message to be displayed in the modal.
}
\description{
This function determines the appropriate modal message based on the
current Git status regarding how many commits the local repository is
ahead or behind the remote repository. It provides instructions for
the user to either push local changes, pull remote updates, or both.
}
