% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{find_root_directory}
\alias{find_root_directory}
\title{Find Root Directory}
\usage{
find_root_directory()
}
\value{
Sets wd to the path of the project root.
}
\description{
This function searches for the root directory of a Git repository by looking for .Rproj file.
}
