% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_files_tree_list}
\alias{get_files_tree_list}
\title{Get a Tree List of Files and Directories}
\usage{
get_files_tree_list(path)
}
\arguments{
\item{path}{A character string specifying the path to retrieve the file tree from.}
}
\value{
A list representing the tree structure of files and directories. Directories are
represented as lists with the class "folder" and files are represented as empty strings
with the class "file".
}
\description{
This function recursively retrieves a tree list of files and directories from a given path.
}
