% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_selected_items}
\alias{get_selected_items}
\title{Get Selected Items from a Tree Structure}
\usage{
get_selected_items(tree, parent = "")
}
\arguments{
\item{tree}{A list representing the tree structure, where each node can have the
attribute "stselected" to indicate selection.}

\item{parent}{A character string specifying the parent path. Defaults to an empty string.}
}
\value{
A list of selected items from the tree structure. The structure of the returned
list mirrors the input tree, with selected items indicated.
}
\description{
This function recursively retrieves selected items from a given tree structure.
Each node in the tree can have an attribute "stselected" which, if TRUE, indicates
that the node is selected.
}
