% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_05_archive.R
\name{ghqc_archive_app}
\alias{ghqc_archive_app}
\title{Archive Files}
\usage{
ghqc_archive_app()
}
\value{
Starts a Shiny app and does not return any value.
}
\description{
This function starts a Shiny application for archiving files.
}
