% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_01_assign.R
\name{ghqc_assign_app}
\alias{ghqc_assign_app}
\title{Assign file(s) for QC}
\usage{
ghqc_assign_app()
}
\value{
Starts a Shiny app and does not return any value.
}
\description{
This function provides an interface to assign one or more files for QC in the form of a GitHub Issue(s) within a
GitHub Milestone, with options to assign a repository collaborator as the QCer and/or generate a checklist
of suggested review tasks during QC.
}
