% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_03_record.R
\name{ghqc_record_app}
\alias{ghqc_record_app}
\title{Generate a QC Record for one or more Milestones}
\usage{
ghqc_record_app()
}
\value{
Starts a Shiny app and does not return any value.
}
\description{
This function allows the user to generate a QC Record for one or more Milestones created with the assign app.
}
