% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_02_resolve.R
\name{ghqc_resolve_app}
\alias{ghqc_resolve_app}
\title{Comment in an Issue to display file changes during QC}
\usage{
ghqc_resolve_app()
}
\value{
Starts a Shiny app and does not return any value.
}
\description{
This function allows a user to insert a comment into a ghqc GitHub Issue that displays changes
in the version control information for the Issue’s corresponding file. By default, the comment
displays both the original and current commits and hashes for the file. These versions are
selected by the user. The comment can optionally display the file difference (“diff”) between
the current and previous versions. These changes will likely be implementations of QC feedback.
}
