% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_repo.R
\name{ghqc_set_info_repo}
\alias{ghqc_set_info_repo}
\title{set the repo that stores the ghqc info}
\usage{
ghqc_set_info_repo(
  repo_path = file.path("~/.local/share/ghqc", info_repo_name())
)
}
\arguments{
\item{repo_path}{path to the git repo storing ghqc information}
}
\description{
set the repo that stores the ghqc info
}
