% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_04_status.R
\name{ghqc_status_app}
\alias{ghqc_status_app}
\title{Status QC file(s)}
\usage{
ghqc_status_app(milestones = NULL)
}
\value{
Starts a Shiny app and does not return any value.
}
\description{
This function provides an interface to status QC files
}
