% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{isolate_rendered_list}
\alias{isolate_rendered_list}
\title{Update Selectize Inputs for Checklists and Assignees}
\usage{
isolate_rendered_list(input, session, items)
}
\arguments{
\item{input}{A reactive list of inputs from a Shiny session.}

\item{session}{A server-side representation of a Shiny session.}

\item{items}{A list representing the selected items.}
}
\value{
None The function performs operations on UI elements and does not return
any value.
}
\description{
This function updates selectize inputs for both checklist and assignee
selections within a Shiny application, handling conditions where assignee
options may vary in number. It specifically preserves the current user
selections for checklists, and automatically selects the sole assignee if
only one is available, or maintains the current selection when more are added.
}
