% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{render_selected_list}
\alias{render_selected_list}
\title{Render a Selected List with Inputs}
\usage{
render_selected_list(
  input,
  ns,
  items = NULL,
  checklist_choices = NULL,
  depth = 0
)
}
\arguments{
\item{input}{A named list from Shiny server function containing inputs from the Shiny application.}

\item{ns}{A namespace function used for handling Shiny modules.}

\item{items}{A list representing the selected items.}

\item{checklist_choices}{A vector of checklist choices for the selectize input fields.}
}
\description{
This function generates a nested list of selected items with associated input fields
for checklists and assignees. It creates a hierarchical representation of items with
file extensions, providing selectize input fields for each file.
}
