% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_04_status.R
\name{ghqc_status_app}
\alias{ghqc_status_app}
\title{Status QC file(s)}
\usage{
ghqc_status_app(milestones = NULL)
}
\arguments{
\item{milestones}{The default set of one or more milestones to render in the QC Status table}
}
\value{
Starts a Shiny app and does not return any value.
}
\description{
View the status of one or more ghqc Issues and post QC Notification(s) and QC Approval(s)
}
